requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
black==19.10b0
bootstrap4==0.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-easy-pdf==0.1.1
django-filter==2.3.0
django-heroku==0.3.1
django-tables2==2.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.8.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.3
PyPDF2==1.26.0
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
regex==2020.6.8
reportlab==3.5.44
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
